Congo Ebola Incidence data from the 2018 Ebola epidemic in the Democratic Republic of Congo.

Congo Ebola Incidence data from the 2018 Ebola epidemic in the Democratic Republic of Congo.

Usage

Congo.EbolaIncidence

Format

Congo.EbolaIncidence

SitReptNo.

An integer labelling the report from the reporting agency.

Date

The date of the report in YYYY-MM-DD format.

Further columns are named for the location in which the integer number of infections ocurred.

  # A tibble: 75 × 30
     SitReptNo. Date                Alimbongo  Beni Biena Butembo  Goma Kalunguta Katwa Kayna Kyondo Lubero Mabalako
          <dbl> <dttm>                  <dbl> <dbl> <dbl>   <dbl> <dbl>     <dbl> <dbl> <dbl>  <dbl>  <dbl>    <dbl>
   1          1 2018-08-05 00:00:00         0     3     0       2     0         0     0     0      0      0       34
   2          2 2018-08-12 00:00:00         0     2     0       0     0         0     0     0      0      0       11
   3          3 2018-08-20 00:00:00         0     1     0       0     0         0     0     0      0      0       37
   4          4 2018-08-26 00:00:00         0     5     0       0     0         0     0     0      0      0        3
   5          5 2018-09-02 00:00:00         0     8     0       0     0         1     0     0      0      0        1
   6          6 2018-09-09 00:00:00         0     5     0       2     0         0     0     0      0      0        1
   7          7 2018-09-16 00:00:00         0     5     0       3     0         0     0     0      0      0        2
   8          8 2018-09-23 00:00:00         0     4     0       1     0         0     0     0      0      0        1
   9          9 2018-10-02 00:00:00         0    10     0       1     0         0     0     0      0      0        0
  10         10 2018-10-07 00:00:00         0    14     0       4     0         0     0     0      0      0        1
  # ℹ 65 more rows
  # ℹ 17 more variables: Manguredjipa <dbl>, Masereka <dbl>, Musienene <dbl>, Mutwanga <dbl>, Nyiragongo <dbl>,
  #   Oicha <dbl>, Pinga <dbl>, Vuhovi <dbl>, Ariwara <dbl>, Bunia <dbl>, Komanda <dbl>, Lolwa <dbl>, Mambasa <dbl>,
  #   Mandima <dbl>, Nyakunde <dbl>, Rwampara <dbl>, Tchomia <dbl>
  # ℹ Use `print(n = ...)` to see more rows