SVEIRD compartmental model with optional Bayesian data assimilation

Run a SVEIRD compartmental model of an epidemic, optionally using Bayesian data assimilation.

Usage

SVEIRD.BayesianDataAssimilation(
  alpha,
  beta,
  gamma,
  sigma,
  delta,
  n.days,
  seedData,
  neighbourhood.order = 0,
  lambda,
  layers,
  aggregationFactor,
  startDate,
  countryCodeISO3C,
  incidenceData = NULL,
  deathData = NULL,
  simulationIsDeterministic = TRUE,
  dataAssimilationEnabled = FALSE,
  healthZoneCoordinates,
  variableCovarianceFunction,
  Q.backgroundErrorStandardDeviation,
  Q.characteristicCorrelationLength,
  neighbourhood,
  psi.diagonal,
  callback = `{`
)

Arguments

alpha

The rate of vaccination (per day)

beta

The rate of exposure (per day)

gamma

The rate of becoming infectious (per day)

sigma

The rate of recovery (per day)

delta

The fatality rate (per day)

n.days

The number of days the simulation will run for, beginning from the startDate.

seedData

a dataframe, as described in initialInfections.fourCities.

neighbourhood.order

The order of the queen's neighbourhood over which to distribute the Exposed and Infected intial state variable; zero is no neighbourhood, and corresponds to only the same cell where the queen already is. Higher orders follow the simple formula (2 * x + 1)^2 to determine the number of cells over which to evenly disperse the exposed and infected state variables. All other state variables are placed directly in the grid cell corresponding to the health zone.

lambda

the average dialy movement distance of an individual (in kilometers).

layers

The SpatRaster object with SVEIRD compartment layers, and a layer classifying habitation. Created with the getSVEIRD.SpatRaster function.

aggregationFactor

The number of adjacent cells in any one direction to aggregate into a single cell. The aggregation factor must be the same as that used to generate the SpatRaster for layers.

startDate

The date (in YYYY-MM-DD format) the simulation begins.

countryCodeISO3C

The ISO three character code for a recognized country.

incidenceData

A "situation report" dataframe. The first column provides the date of the officially reported, observed incidence of the disease, in ISO format (YYYY-MM-DD). MAYBE TODO: enforce the startDate parameter to be one week prior to the first observed data?

    Date    Beni  Biena  Butembo  Goma  Kalunguta
    05-Aug  3     0      2        0     0
    12-Aug  2     0      0        0     0
    20-Aug  1     0      0        0     0
    26-Aug  5     0      0        0     0
    02-Sep  8     0      0        0     1
    09-Sep  5     0      2        0     0
    16-Sep  5     0      3        0     0
    23-Sep  4     0      1        0     0
    02-Oct  10    0      1        0     0
    07-Oct  14    0      4        0     0
    15-Oct  28    0      2        0     1
    21-Oct  19    0      3        0     0
    28-Oct  28    0      8        0     0
    04-Nov  16    0      6        0     1
    11-Nov  14    0      0        0     21
  

deathData

Data of the same format as incidenceData, but observations represent deaths, not infections.

simulationIsDeterministic

Whether stochasticity is enabled or not; if the simulation is deterministic then no stochastic processes are used and the simulation is entirely deterministic. Either TRUE or FALSE.

dataAssimilationEnabled

Whether Bayesian data assimilation will be used for state reporting data.

healthZoneCoordinates

The coordinates of health zones in the country of interest, TODO: describe the use of the data so users understand why they must include it; e.g. (which will be used to group and summarize the compartmental model data at the end of the simmulation.)

    HealthZone    Latitude    Longitude
    Alimbongo     -0.365515   29.1911818
    Beni          0.49113     29.47306
    Biena         0.57923     29.115633
    Butembo       0.140692    29.335014
    Goma          -1.658271   29.220132
    Kalunguta     0.323085    29.354677
    Katwa         0.116985    29.411838
    Kayna         -0.603936   29.174066
    Kyondo        -0.005622   29.408813
    Lubero        -0.15633    29.24057
    Mabalako      0.461257    29.210687
    Manguredjipa  0.353433    28.765479
    Masereka      -0.133333   29.333333
    Musienene     0.04022     29.26246
    Mutwanga      0.32893     29.74576
    Nyiragongo    -1.516667   29.25
    Oicha         0.698681    29.518834
    Pinga         -0.9830504  28.687911
    Vuhovi        0.1416      29.4075
    Ariwara       3.136873    30.706615
    Bunia         1.566667    30.25
    Komanda       1.367496    29.774322
    Lolwa         1.352969    29.498455
    Mambasa       1.359731    29.029226
    Mandima       1.35551     29.08173
    Nyakunde      1.431271    30.029283
    Rwampara      1.4053      30.3449
    Tchomia       1.4412      30.4845
  

variableCovarianceFunction

Passed directly to Q.forecastErrorCov() to generate a Q matrix.

Q.backgroundErrorStandardDeviation

TODO

Q.characteristicCorrelationLength

TODO

neighbourhood

TODO

psi.diagonal

TODO

callback

a callback function to run, with no arguments, which will be called every time the main loop of the simulation iterates.

Value

a summaryTable dataframe for the simulation, showing changes in the compartment values over time, the daily values, and cumulative values.

Details

TODO: DETAILS of the function.

Author

Bryce Carson

Ashok Krishnmaurthy

Michael Myer

Examples

subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
data("healthZonesCongo", package = "spatialEpisim.foundation")
data("initialInfections.fourCities", package = "spatialEpisim.foundation")
data("Congo.EbolaIncidence", package = "spatialEpisim.foundation")
SVEIRD.BayesianDataAssimilation(
  ## Parameters
  alpha = 3.5e-5,
  beta = 7e-3,
  gamma = 1/7,
  sigma = 1/36,
  delta = 2/36,
  lambda = 15,
  ## Model runtime
  n.days = 31, # a month, permitting three assimilations of observed data
  ## Model data
  seedData = initialInfections.fourCities,
  neighbourhood.order = 1,
  layers = getSVEIRD.SpatRaster(subregionsSpatVector,
                                susceptibleSpatRaster,
                                aggregationFactor = 35),
  aggregationFactor = 35,
  startDate = "2018-08-05",
  countryCodeISO3C = "COD",
  incidenceData = Congo.EbolaIncidence,
  ## Model options
  simulationIsDeterministic = TRUE,
  dataAssimilationEnabled = TRUE,
  healthZoneCoordinates = healthZonesCongo,
  variableCovarianceFunction = "DBD",
  ## Special parameters
  Q.backgroundErrorStandardDeviation = 0.55,
  Q.characteristicCorrelationLength = 6.75e-1,
  neighbourhood = 3,
  psi.diagonal = 1e-3
)
#> Susceptible before seeding = 126108
#> Susceptible after seeding = 6461
#> Warning: Raster aggregation factor is too high to differentiate between two (or more) health zones (they correspond to the same grid cell).
#> Error: [focal] not a meanigful window