Download WorldPop 2020 data

Download 1 km aggregated UN-adjusted population count spatial data from 2020 through the WorldPop servers.

Usage

downloadWorldPopData(
  countryCodeISO3C,
  folder = getOption("spatialEpisim.foundation.datapath")
)

Arguments

countryCodeISO3C

The uppercase ISO three character code a recognized country.

folder

The destination folder the downloaded data will be stored in

Value

An absolute path where the data was downloaded, or the path at which the file already existed

Details

This function only downloads the file if necessary (if it doesn't exist at path).

ppp is the code used by WorldPop to mean population estimate for population counts; similar data, but which grids population density is coded pd rather than ppp. 1km means 1 km aggregation of the data. The description WorldPop gives for this data is "individual countries 2000-2020 UN adjusted aggregated to 1km resolution using 100m resolution population count datasets". See: https://hub.worldpop.org/doi/10.5258/SOTON/WP00671.

Author

Bryce Carson

Examples

downloadWorldPopData(countryCodeISO3C = "COD", folder = tempdir())
#> [1] "/home/runner/spatialEpisim.foundation.data/cod_ppp_2020_1km_Aggregated_UNadj.tif"