Create a SpatRaster of SVEIRD model compartment raster data

Create a list of SpatRaster objects, one for each component in an SVEIRD epidemic model.

Usage

getSVEIRD.SpatRaster(subregions, susceptible, aggregationFactor = NULL)

Arguments

subregions

a SpatVector object of subregions used to crop the SVEIRD SpatRaster

susceptible

a RasterLayer, pre-aggregated if that is wished.

aggregationFactor

the number of cells in any direction to aggregate together into one, in the susceptible ratser, after masking with the subregions vector, before creating the list

Value

a SpatRaster, with layers for the SVERID components

Details

The SpatRaster objects for the VEIRD components are empty, the input SpatRaster is the only layer with non-zero values (if any existed before).

Author

Bryce Carson

Michael Myer

Ashok Krishnmaurthy

Examples

subregionsSpatVector <- terra::vect(
  system.file(
    "extdata",
    ## COD: Nord-Kivu and Ituri (Democratic Republic of Congo)
    "subregionsSpatVector",
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
susceptibleSpatRaster <- terra::rast(
  system.file(
    "extdata",
    "susceptibleSpatRaster.tif", # Congo population
    package = "spatialEpisim.foundation",
    mustWork = TRUE
  )
)
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster,
                     aggregationFactor = 10)
#> class       : SpatRaster 
#> dimensions  : 69, 49, 6  (nrow, ncol, nlyr)
#> resolution  : 0.08333333, 0.08333333  (x, y)
#> extent      : 27.22375, 31.30708, -2.075416, 3.674584  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#> source(s)   : memory
#> names       : Susceptible, Vaccinated, Exposed, Infected, Recovered, Dead 
#> min values  :           0,        NaN,     NaN,      NaN,       NaN,  NaN 
#> max values  :         565,        NaN,     NaN,      NaN,       NaN,  NaN 

## Omitting the aggregation factor argument will prevent aggregation. An
## aggregation factor of zero or one is meaningless and will produce an
## error.
getSVEIRD.SpatRaster(subregionsSpatVector, susceptibleSpatRaster)
#> class       : SpatRaster 
#> dimensions  : 689, 487, 6  (nrow, ncol, nlyr)
#> resolution  : 0.008333333, 0.008333333  (x, y)
#> extent      : 27.22375, 31.28208, -2.067083, 3.674584  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#> source(s)   : memory
#> varnames    : susceptibleSpatRaster 
#>               susceptibleSpatRaster 
#> names       : Susceptible, Vaccinated, Exposed, Infected, Recovered, Dead 
#> min values  :           0,        NaN,     NaN,      NaN,       NaN,  NaN 
#> max values  :           6,        NaN,     NaN,      NaN,       NaN,  NaN