Reporting health zones of the Democratic Republic of Congo

Reporting health zones of the Democratic Republic of Congo

Usage

healthZonesCongo

Format

healthZonesCongo

HealthZone

The name of the Congolese health zone

Latitude

The geographical latitude of the health zone

Longitude

The geographical longitude of the health zone

       HealthZone   Latitude Longitude
  1     Alimbongo -0.3655150  29.19118
  2          Beni  0.4911300  29.47306
  3         Biena  0.5792300  29.11563
  4       Butembo  0.1406920  29.33501
  5          Goma -1.6582710  29.22013