Retrieve a SpatRaster of level 1 boundaries from local or remote disk

Returns a terra::SpatVector for the requested country on demand, either retrieving the data from disk if it has been downloaded before, or downloading the SpatVector to disk at the location determined by geodata::geodata_path().

Usage

lvl1AdminBorders(countryCodeISO3C, folder)

Arguments

countryCodeISO3C

The uppercase ISO three character code a recognized country.

folder

The path where GADM data should be found or stored; passed on to geodata::gadma.

Value

SpatVector

Details

The level one boundaries are the least granular (geographically largest) administrative boundaries that countries create to subdivide themselves.

The default resolution is one kilometer, meaning points in the vector are no closer to each other than one kilometer, therefore details smaller than one kilometer cannot be differentiated using this data.

Author

Bryce Carson

Examples

library(geodata)
geodata_path_backup <- geodata_path()
options(geodata_default_path = getOption("spatialEpisim.foundation.datapath"))
lvl1AdminBorders("COD")
#>  class       : SpatVector 
#>  geometry    : polygons 
#>  dimensions  : 26, 10  (geometries, attributes)
#>  extent      : 12.20663, 31.30572, -13.45568, 5.386098  (xmin, xmax, ymin, ymax)
#>  coord. ref. : +proj=longlat +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +no_defs 
#>  names       : GID_0          NAME_0   GID_1       NAME_1 VARNAME_1 NL_NAME_1
#>  type        : <chr>           <chr>   <chr>        <chr>     <chr>     <chr>
#>  values      :   COD Democratic Rep~ COD.1_1     Bas-Uélé        NA        NA
#>                  COD Democratic Rep~ COD.2_1     Équateur        NA        NA
#>                  COD Democratic Rep~ COD.3_1 Haut-Katanga        NA        NA
#>    TYPE_1 ENGTYPE_1  CC_1 HASC_1
#>     <chr>     <chr> <chr>  <chr>
#>  Province  Province    NA  CD.BU
#>  Province  Province    NA  CD.ET
#>  Province  Province    NA  CD.HK
options(geodata_default_path = geodata_path_backup)

if (FALSE) { # \dontrun{
## It's recommended to set geodata_default_path in your .Rprofile rather
## than explicitly provide the path.
lvl1AdminBorders("COD", file.path("data", "gadm"))
} # }